+++ 1.21/Fabric/src/main/java/fuzs/forgeconfigapiport/fabric/api/forge/v4/ForgeModConfigEvents.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 public final class ForgeModConfigEvents {
 
     private ForgeModConfigEvents() {
-
+        // NO-OP
     }
 
     /**