+++ 1.21/Forge/src/main/java/fuzs/forgeconfigapiport/forge/api/neoforge/v4/NeoForgeConfigRegistry.java
@@ -1,14 +1,16 @@
 package fuzs.forgeconfigapiport.forge.api.neoforge.v4;
 
 import fuzs.forgeconfigapiport.forge.impl.neoforge.NeoForgeConfigRegistryImpl;
-import net.minecraftforge.fml.ModLoadingContext;
+import net.minecraftforge.fml.ModContainer;
 import net.minecraftforge.fml.config.ModConfig;
 import net.neoforged.fml.config.IConfigSpec;
 
 /**
- * Registry for adding configs built using {@link net.neoforged.neoforge.common.ModConfigSpec}. Serves as a bridge between NeoForge's and Forge's config systems.
- *
- * <p>Methods must be called only during {@link net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent}, as would be done for Forge's configs. Do not call during any of the setup events.
+ * Registry for adding configs built using {@link net.neoforged.neoforge.common.ModConfigSpec}. Serves as a bridge
+ * between NeoForge's and Forge's config systems.
+ * <p>
+ * Methods must be called only during {@link net.neoforged.fml.event.lifecycle.FMLConstructModEvent}, as would be done
+ * for Forge's configs. Do not call during any of the setup events.
  */
 public interface NeoForgeConfigRegistry {
     NeoForgeConfigRegistry INSTANCE = new NeoForgeConfigRegistryImpl();
@@ -16,13 +18,11 @@ public interface NeoForgeConfigRegistry {
     /**
      * Register a new mod config.
      *
-     * @param type  type of this mod config
-     * @param spec  the built config spec
+     * @param type type of this mod config
+     * @param spec the built config spec
      * @return the {@link ModConfig} instance
      */
-    default ModConfig register(ModConfig.Type type, IConfigSpec<?> spec) {
-        return this.register(ModLoadingContext.get().getActiveNamespace(), type, spec);
-    }
+    ModConfig register(ModConfig.Type type, IConfigSpec spec);
 
     /**
      * Register a new mod config.
@@ -32,7 +32,17 @@ public interface NeoForgeConfigRegistry {
      * @param spec  the built config spec
      * @return the {@link ModConfig} instance
      */
-    ModConfig register(String modId, ModConfig.Type type, IConfigSpec<?> spec);
+    ModConfig register(String modId, ModConfig.Type type, IConfigSpec spec);
+
+    /**
+     * Register a new mod config.
+     *
+     * @param modContainer mod container to register config for
+     * @param type         type of this mod config
+     * @param spec         the built config spec
+     * @return the {@link ModConfig} instance
+     */
+    ModConfig register(ModContainer modContainer, ModConfig.Type type, IConfigSpec spec);
 
     /**
      * Register a new mod config.
@@ -42,9 +52,7 @@ public interface NeoForgeConfigRegistry {
      * @param fileName file name to use instead of default
      * @return the {@link ModConfig} instance
      */
-    default ModConfig register(ModConfig.Type type, IConfigSpec<?> spec, String fileName) {
-        return this.register(ModLoadingContext.get().getActiveNamespace(), type, spec, fileName);
-    }
+    ModConfig register(ModConfig.Type type, IConfigSpec spec, String fileName);
 
     /**
      * Register a new mod config.
@@ -55,5 +63,16 @@ public interface NeoForgeConfigRegistry {
      * @param fileName file name to use instead of default
      * @return the {@link ModConfig} instance
      */
-    ModConfig register(String modId, ModConfig.Type type, IConfigSpec<?> spec, String fileName);
+    ModConfig register(String modId, ModConfig.Type type, IConfigSpec spec, String fileName);
+
+    /**
+     * Register a new mod config.
+     *
+     * @param modContainer mod container to register config for
+     * @param type         type of this mod config
+     * @param spec         the built config spec
+     * @param fileName     file name to use instead of default
+     * @return the {@link ModConfig} instance
+     */
+    ModConfig register(ModContainer modContainer, ModConfig.Type type, IConfigSpec spec, String fileName);
 }