+++ 1.21/CHANGELOG.md
@@ -2,9 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
+## [v21.0.1-1.21] - 2024-07-01
+- Updated to Night Config v3.8.0
+- Update for NeoForge v21.0.82
+### Changed
+- Remove internal Forge classes from the Fabric publication, NeoForge's system is used to benefit from the concurrency safety it provides
 
 ## [v21.0.1-1.21] - 2024-07-01
-- Updated for Night Config v3.7+
+- Updated for Night Config v3.7.0
 ### Fixed
 - Common publications no longer try to pull Fabric Loader